Moved CreateEmptyTestTable out dtestutils
